+++ model-based-configurable-nat/src/main/java/org/onosproject/model/based/configurable/nat/StateListenerNew.java
@@ -162,10 +162,10 @@ public class StateListenerNew extends Thread{
                 IpAddress value = IpAddress.valueOf(json);
                 return value;
             }
-            if(type == Long.class){
-//		log.info("long recevide: " + json);
-                Long value = Long.parseLong(json);
-//		log.info("long converted: " + value);
+            if(type == Integer.class){
+//		log.info("int recevide: " + json);
+                Integer value = Integer.parseInt(json);
+//		log.info("int converted: " + value);
                 return value;
             }
             if(type == PortNumber.class){
@@ -227,7 +227,7 @@ public class StateListenerNew extends Thread{
         if(type.equals("int32"))
             return Integer.parseInt(value.toString());
         if(type.equals("inet:port-number"))
-            return Long.parseLong(value.toString());
+            return Integer.parseInt(value.toString());
         return value.toString();
     }
     
@@ -1968,7 +1968,7 @@ log.info("***STATIC LIST INDEXES FOUND IN THE MAPPING FILE***");
                             ObjectNode entry = (ObjectNode)tableEntries.next();
                             log.info("entry "+entry);
                             Ip4Address srcIp=null, dstIp=null, translatedIp=null;
-                            Long srcPort=null, dstPort=null, translatedPort=null;
+                            Integer srcPort=null, dstPort=null, translatedPort=null;
                             String connectionState = null;
 		            byte protocol = 0;
 
@@ -1981,11 +1981,11 @@ log.info("***STATIC LIST INDEXES FOUND IN THE MAPPING FILE***");
                             if(entry.has("translated_address"))
                                 translatedIp = Ip4Address.valueOf(entry.get("translated_address").textValue()); 
                             if(entry.has("src_port"))
-                                srcPort = (long)entry.get("src_port").asLong();
+                                srcPort = (int)entry.get("src_port").asInt();
                             if(entry.has("dst_port"))
-                                dstPort = (long)entry.get("dst_port").asLong();
+                                dstPort = (int)entry.get("dst_port").asInt();
                             if(entry.has("translated_port"))
-                                translatedPort = (long)entry.get("translated_port").asLong();
+                                translatedPort = (int)entry.get("translated_port").asInt();
                             if(entry.has("tcp_state"))
                                 connectionState = entry.get("tcp_state").textValue();	
 			    //byte[] byteArray = entry.get("protocol").toString().getBytes("utf-8");
